+++ __test__/setup.ts
@@ -1,4 +1,0 @@
-// suppress all those red colored errors
-window.console.error = (e: Error) => {
-  e instanceof Error && console.log(e.message)
-};