+++ src/util.ts
@@ -1,14 +1,12 @@
-const parseJSON = <T extends any>(value: string): T | string => {
-  try {
-    return JSON.parse(value);
-  } catch (e)  /*istanbul ignore next*/ {
-    console.error(e);
+const parseJSON = <T extends any>(value: string | null): T | string | null => {
+  if (value === null) {
     return value;
   }
+  return JSON.parse(value);
 };
 
 const arrayify = (item: any | []): any[] => {
   return item instanceof Array ? item : [item];
 };
 
-export { parseJSON, arrayify }
+export {parseJSON, arrayify}